fix communication bugs of mrvl wifi, project support wifi,ethernet,at modules....
