+++ components/connectivity/agent_tiny/examples/agent_tiny_demo.c
@@ -39,7 +39,8 @@
 //#define DEFAULT_SERVER_IPV4 "139.159.209.89"/*Huawei */
 //#define DEFAULT_SERVER_IPV4 "192.168.0.116"/*Huawei */
 //#define DEFAULT_SERVER_IPV4 "192.168.1.106"/*sjn */
-#define DEFAULT_SERVER_IPV4 "192.168.1.113" /*czr*/
+//#define DEFAULT_SERVER_IPV4 "192.168.1.113" /*czr*/
+#define DEFAULT_SERVER_IPV4 "192.168.1.100" /*mj*/
 
 #define LWM2M_LIFE_TIME     50000
