--- targets/Cloud_wifi_STM32F429ZI_NUCLEO/Inc/wifi_sdio_demo.h
@@ -1,49 +1,0 @@
-/******************************************************************************
-
-  Copyright (c) <2016-2018>, <Huawei Technologies Co., Ltd>
-
- ******************************************************************************
-  File Name     : wifi_sdio_demo.h
-  Version       : Initial Draft
-  Author        : MAMINJIE
-  Created       : 2018/5/26
-  Last Modified :
-  Description   : wifi_sdio_demo.c header file
-  Function List :
-  History       :
-  1.Date        : 2018/5/26
-    Author      : MAMINJIE
-    Modification: Created file
-
-******************************************************************************/
-/* Define to prevent recursive inclusion ------------------------------------*/
-#ifndef __WIFI_SDIO_DEMO_H__
-#define __WIFI_SDIO_DEMO_H__
-
-
-#ifdef __cplusplus
-#if __cplusplus
-extern "C" {
-#endif
-#endif /* __cplusplus */
-
-/* Includes -----------------------------------------------------------------*/
-#include "wifi.h"
-
-/* Defines ------------------------------------------------------------------*/
-/* Macros -------------------------------------------------------------------*/
-/* Typedefs -----------------------------------------------------------------*/
-/* Extern variables ---------------------------------------------------------*/
-/* Functions API ------------------------------------------------------------*/
-void wifi_sdio_demo(void);
-
-
-
-#ifdef __cplusplus
-#if __cplusplus
-}
-#endif
-#endif /* __cplusplus */
-
-
-#endif /* __WIFI_SDIO_DEMO_H__ */
